(Fully) local integration tests
This commit introduces a way to run the integration tests in a purely
local fashion, using `fakes3` (a Ruby gem) rather than using an actual
S3 bucket.

In order to accomplish this a few configuration options needs to be set
(AWS keys and S3 bucket), and I’ve included “defaults” that will work on
Travis (in separate config. files).
One also needs to add an entry to `/etc/hosts` (or equivalent), that
points `test-bucket.localhost` to `127.0.0.1`.

I’ve added a rather naive function, `check_for_native_libs`, that checks
whether `HADOOP_NATIVE_LIB_PATH` actually contains any `*.so` files, and
if it doesn’t it’ll ignore the “compressed tests” for sequence files, as
it's not possible to run them without the native libs, and it's pretty
darn difficult to build them on Mac OS X, not even sure if it's
supported at all.

Furthermore there’s a Makefile that simplifies running the integration
tests, a simple `make integration` is all that’s needed. It’ll build,
package and extract Secor to a directory in `/tmp` and the run the tests.

Perhaps there should be a section in the README dedicated to how the
test setup works, and how to run the tests.