Different log message for new partition
Currently when Secor starts consuming a “new” partition it logs a
warning that the offset for a given partition changed from -2 to
*something*, and I’d say that this particular scenario is expected and
shouldn’t generate a warning, but rather an info message that
consumption of a new partition has started.

For future reference, -2 means to start from the earliest available
offset and -1 means to start from the latest available offset, if I
remember correctly.