Adding a shutdown hook to clean up the local files before exiting.
When secor process died, often we found there are a large set of local data files left on the disk.  Although we have a cleanup thread to clean files older than 24 hours, but sometimes this is already too late.

Add a shutdown hook to clean up files on process exit.

Changed a log message from debug to warning to help the debug.