Support timestamp fields past the first field in ThriftMessageParser
This commit adds configurable functionality to ThriftmessageParser for reading a timestamp in fields other than field ID 1. Additionally timestamps in i32 fields re now supported as well as i64.

Includes new properties:
message.timestamp.id: (default 1) The ID or Index for the timestamp field name
message.timestamp.type: (default i64) The type of the field. Rather than the previously assumed i64 the thrift parser can now read i32. (just in case you want to subject yourself to the 2038 problem). This will fall back to i64 if the value is unrecognized.

Also added a test suite for the ThriftMessageParser.

N.B. Benchmarks suggest that there is an immediate performance penalty just for reading past the first field, regardless of the size of the message.