Remove 1p,custom.properties and move all its content to 1p-secor-service-dev.properties and 1p-secor-service.properties.
It is because we should not hard-coded the aws keys, kafka, and zookeeper. Instead, we should get them from
archaius, the 1p*.properties are uploaded to Eiddo and the application should get those configs from there.
Note that Secor has it own set of properties, we merged the configs read from them and Eiddo in SecorConfig.