All for git or env var to drive build.properties generation
This allows you to set an env var (SOURCE_VERSION) instead of relying on
git command line to be present to generate the build.properties file.
Very useful if you deploy to an environment where git may not be
present, or the checkout may have it's ties back to the git repo from
whence it came severed, as is the case when deploying to Heroku.