Clean up the log4j properties because it is never used.
kafka uses log4j in its codes (Logging.scala) So I have to put log4j
back in.

I think the reason why logback worked in older instance is because of
multibinding on slf4j, when the logback binding being loaded first, then
logback is used. If not, then other binding is used.