Fix root project name for "1p-secor-service-v2"
Currently, we have two secor running, "1psecorservice" and
"1psecorservicev2", each of them writes to a different bucket. However,
we miseed the point that both secor are using the same kafka group name
used in zk which makes both secor reading the offsets from the same
place, which is not correct.