Allow to set -Dconfig flag via env CONFIG_FILE on startup.
This allows us refer to another CONFIG_FILE during startup.  I've kept the
default which refer to `secor.prod.partition.properties` if not set.

In kubernetes we can simply control which config is used by volume mounting the
correct config map, but this keeps backwards compatibility and provides the
possibilities to:

 * All environments in one config map
 * A given environment in one config map.