Add note about JRE8 requirement
README should contain this note as source code uses JRE8 specific language
features which is not available in < JRE8.

Probably good idea to just leave a note about JRE9 and JRE10 is untested as
well due to nature that JRE9 came with module support which I assume haven't
been verified that it works well with. Requires run-thru on dependencies to
ensure everything works as it should if you want to make this statement.